backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-core==2.41.5
starlette==0.37.2
typing-extensions==4.15.0
annotated-types==0.7.0
typing-inspection==0.4.2
anyio==4.13.0
idna==3.11
click==8.3.1
h11==0.16.0
dnspython==2.8.0
